+++ bootstrapper/cards/cardsNg2.html
@@ -29,6 +29,7 @@
 <div>
 	<label>Card container:</label>
 	<rlCardContainer [builder]="builder">
+		<div *rlColumnContent="let value; name 'value'">#{{value}}</div>
 		<div *rlCardContent="let myItem">
 			Name: {{myItem.name}}
 			Value: {{myItem.value}}
