+++ source/components/cardContainer/card/headerColumn/headerColumn.html
@@ -1,0 +1,6 @@
+<div [ngClass]="sizeClass" [title]="column?.description">
+	<div *ngIf="!columnTemplate">{{value}}</div>
+	<div *ngIf="columnTemplate">
+		<div [ngTemplateOutlet]="columnTemplate.template" [ngOutletContext]="context"></div>
+	</div>
+</div>
