+++ source/components/cardContainer/card/headerColumn/headerColumn.module.ts
@@ -2,8 +2,8 @@ import * as angular from 'angular';
 
 import { services, downgrade } from 'typescript-angular-utilities';
 
-import { directiveName, headerColumn, controllerName, HeaderColumnController } from './headerColumn';
-import { sizeForBreakpointsName, sizeForBreakpoints } from './sizeForBreakpoints';
+import { directiveName, headerColumn, controllerName, HeaderColumnController } from './headerColumn.ng1';
+import { sizeForBreakpointsName, sizeForBreakpoints } from './sizeForBreakpoints.ng1';
 
 export var moduleName: string = 'rl.ui.components.cardContainer.card.headerColumn';
